LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-report-only-expected.txt

CONSOLE MESSAGE: [Report Only] This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: [Report Only] This requires a TrustedScriptURL value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: [Report Only] This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Hello
CONSOLE MESSAGE: [Report Only] This requires a TrustedScriptURL value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
abc

PASS Trusted Type violation report-only: assign string to html
PASS Trusted Type violation report-only: assign string to script.src
PASS Trusted Type violation report-only: assign string to script content
PASS Trusted Type violation report: check report contents
